requirements.txt
django
psycopg2
python-decouple
djangorestframework
gunicorn
django-filter
